Add files via upload
The file here took 1 year of data on a 4 hour candle stick chart (11,000+ lines of data) and tested on my strategy to see what the results would be. It then writes that to an excel file for further analysis. This code can be reused with different trading pairs if the civ files have the right column names.